Move code to smaller components
